+++ packages/rax-scrollview/src/utils/wrapDefaultProperties.ts
@@ -7,7 +7,7 @@ const DEFAULT_SCROLL_CALLBACK_THROTTLE = 50;
 export default function wrapDefaultProperties(ScrollView: ForwardRefExoticComponent<ScrollViewProps>): ForwardRefExoticComponent<ScrollViewProps> {
   ScrollView.defaultProps = {
     scrollEventThrottle: DEFAULT_SCROLL_CALLBACK_THROTTLE,
-    onEndReachedThreshold: DEFAULT_END_REACHED_THRESHOLD,
+    endReachedThreshold: DEFAULT_END_REACHED_THRESHOLD,
     showsHorizontalScrollIndicator: true,
     showsVerticalScrollIndicator: true,
     className: 'rax-scrollview'