+++ packages/rax-scrollview/src/weex/index.tsx
@@ -4,7 +4,6 @@ import {
   ForwardRefExoticComponent,
   forwardRef,
   useRef,
-  useState,
   useImperativeHandle
 } from 'rax';
 import View from 'rax-view';
@@ -27,11 +26,10 @@ const ScrollView: ForwardRefExoticComponent<ScrollViewProps> = forwardRef(
       showsHorizontalScrollIndicator,
       showsVerticalScrollIndicator,
       onEndReached,
-      onEndReachedThreshold,
+      endReachedThreshold,
       onScroll,
       children
     } = props;
-    const [loadmoreretry, setLoadmoreretry] = useState(0);
     const scrollerRef = useRef<HTMLDivElement>(null);
     const contentContainerRef = useRef<HTMLDivElement>(null);
     const handleScroll = e => {
@@ -52,8 +50,9 @@ const ScrollView: ForwardRefExoticComponent<ScrollViewProps> = forwardRef(
     };
     useImperativeHandle(ref, () => ({
       _nativeNode: scrollerRef.current,
-      resetScroll() {
-        setLoadmoreretry(loadmoreretry + 1);
+      resetEndReached() {
+        // @ts-ignore
+        scrollerRef.current.resetLoadmore();
       },
       scrollTo(options?: {
         x?: number;
@@ -95,10 +94,10 @@ const ScrollView: ForwardRefExoticComponent<ScrollViewProps> = forwardRef(
     }));
 
     // In weex must be int value
-    onEndReachedThreshold =
-      typeof onEndReachedThreshold === 'string'
-        ? parseInt(onEndReachedThreshold, 10)
-        : onEndReachedThreshold;
+    endReachedThreshold =
+      typeof endReachedThreshold === 'string'
+        ? parseInt(endReachedThreshold, 10)
+        : endReachedThreshold;
     if (style) {
       const childLayoutProps = ['alignItems', 'justifyContent'].filter(
         prop => style[prop] !== undefined
@@ -169,8 +168,7 @@ const ScrollView: ForwardRefExoticComponent<ScrollViewProps> = forwardRef(
         showScrollbar={showsScrollIndicator}
         onLoadmore={onEndReached}
         onScroll={onScroll ? handleScroll : null}
-        loadmoreoffset={onEndReachedThreshold}
-        loadmoreretry={loadmoreretry}
+        loadmoreoffset={endReachedThreshold}
         scrollDirection={horizontal ? 'horizontal' : 'vertical'}
       >
         {refreshContainer}